move codec utils
